Add environment variables for data location and Hive db & table; add task for data ingest
Add new user-configurable environment variables for the locations (filepath + Hive table) that we write to.
Add task to run data ingest, which was previously skipped.
Amended lineage.yml to reflect new default value for Hive table, but when Prototypes support making use of the lineage file we will want to revisit this and programatically edit it to reflect the actual Hive table name the user selected.